pyproject.toml
[project]
name = "simplenote-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
    "simplenote>=2.1.4",
    "typer>=0.24.1",
]

[project.scripts]
sn = "main:app"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
